net: shell: Add ipv4 command
Move "ipv4" command to a separate file.
